Add files via upload
Post-quantum cryptography is a term used in the field of cryptography to describe cryptographic algorithms (typically public-key algorithms) that are thought to be secure against a cryptanalytic attack by a quantum computer. It is also referred to as quantum-proof, quantum-safe, or quantum-resistant.
The issue with existing common algorithms is that their security depends on one of three challenging mathematical problems: the discrete logarithm problem, the discrete logarithm problem with elliptic curves, or the integer factorization problem.
On a quantum computer running Shor's algorithm and with appropriate power, all of these issues may be readily resolved. 
Numerous cryptographers are creating new algorithms in anticipation of the day when quantum computing poses a threat, despite the fact that the experimental quantum computers that are now known to the public lack the processing capability to crack any actual cryptographic scheme.
Since 2006, the PQCrypto conference series has drawn increased interest from academia and business, and more recently, the European Telecommunications Standards Institute (ETSI) and the Institute for Quantum Computing have organised a number of Quantum Safe Cryptography seminars. 
Contrary to the danger that quantum computing poses to existing public-key methods, the majority of existing symmetric cryptographic algorithms and hash functions are thought to be quite safe against assaults by quantum computers.
Increasing the key size can successfully thwart attacks against symmetric cyphers, even though the quantum Grover's technique speeds up these attacks.
Therefore, there is no need for post-quantum symmetric cryptography to deviate considerably from symmetric cryptography as it is now.